Revert "add new templates (#1408)" (#1413)
This reverts commit b9814685e3aed4fed5f8f4c08fd0f8e35940979c.
