--- common/components/Forms/GroupedReceiptForm/templates/classic/index.tsx
@@ -1,114 +1,0 @@
-import { FC } from 'react'
-import dayjs from 'dayjs'
-import GroupedPricesTable from '@components/Forms/GroupedReceiptForm/GroupedPricesTable'
-import { TemplateProps } from '../types'
-import cs from './style.module.scss'
-
-const ClassicTemplate: FC<TemplateProps> = ({
-  data,
-  componentRef,
-  isEnglish,
-  currencyLabel,
-  currency,
-  rows,
-  getQty,
-  subtotal,
-  taxPercent,
-  taxAmount,
-  total,
-  paymentInfoLines,
-  issuedToLines,
-  normalizedBankDetailsLines,
-}) => (
-  <div
-    className={cs.invoiceContainer}
-    ref={componentRef}
-    style={{
-      width: '100%',
-      height: '100%',
-      marginTop: '2em',
-      marginRight: '1.5em',
-      marginLeft: '1.5em',
-    }}
-  >
-    <div className={cs.providerInfo}>
-      <div className={cs.label}>{isEnglish ? 'Provider' : 'Постачальник'}</div>
-      <pre className={cs.preLabel}>
-        {data?.provider?.description?.trim()} <br />
-        <br />
-      </pre>
-    </div>
-
-    <div className={cs.receiverInfo}>
-      <div className={cs.label}>{isEnglish ? 'Recipient' : 'Одержувач'}</div>
-      <pre className={cs.preLabel}>
-        {data?.reciever?.description?.trim()} <br />
-        {data?.reciever?.companyName} <br />
-        {data?.reciever?.adminEmails?.map((email: string) => (
-          <div key={email}>
-            {email} <br />
-          </div>
-        ))}
-      </pre>
-    </div>
-
-    <div className={cs.providerInvoice}>
-      <div className={cs.datecellTitle}>
-        {isEnglish ? 'INVOICE №' : 'РАХУНОК №'} {data.invoiceNumber}
-      </div>
-      <div className={cs.datecellDate}>
-        {isEnglish ? 'dated' : 'Від'} &nbsp;
-        {dayjs(data?.invoiceCreationDate)?.format?.('DD.MM.YYYY')}
-        {isEnglish ? '.' : ' року.'}
-      </div>
-      <div className={cs.datecell}>
-        {isEnglish ? 'Due by' : 'Підлягає сплаті до'} &nbsp;
-        {dayjs(data?.invoiceCreationDate).add(5, 'd').format('DD.MM.YYYY')}
-        {!isEnglish && (
-          <>
-            &nbsp; року
-          </>
-        )}
-      </div>
-    </div>
-
-    <div className={cs.tableSum}>
-      <GroupedPricesTable
-        preview
-        domainId={data?.domain?._id ?? data?.domain}
-        currency={currency}
-        invoices={data?.invoice ?? []}
-      />
-    </div>
-
-    <div className={cs.payTable}>
-      <div className={cs.payFixed}>
-        {isEnglish ? 'Total payment amount:' : 'Загальна сума оплати:'}
-        <div className={cs.payBoldSum}>
-          {(+data?.generalSum || +data?.debit || 0).toFixed(2)}{' '}
-          {currencyLabel}
-        </div>
-      </div>
-
-      <div>
-        {isEnglish ? 'Payment purpose:' : 'Призначення платежу:'}{' '}
-        <strong>
-          {isEnglish
-            ? `Payment for services according to invoice № ${data.invoiceNumber} dated ${dayjs(
-                data?.invoiceCreationDate
-              )?.format?.('DD.MM.YYYY')}`
-            : `Оплата за послуги згідно рахунку № ${data.invoiceNumber} від ${dayjs(
-                data?.invoiceCreationDate
-              )?.format?.('DD.MM.YYYY')}`}
-        </strong>
-      </div>
-
-      <div className={cs.payFixed}>
-        {data?.provider?.description?.split('\n')?.[0] || ''}
-        <div className={cs.lineInner}>________________</div>
-      </div>
-    </div>
-  </div>
-)
-
-export default ClassicTemplate
