+++ common/components/Forms/GroupedReceiptForm/templates/style.module.scss
